Move GroupSubStatus to sentry.typing
