+++ src/sentry/tasks/clear_expired_snoozes.py
@@ -1,10 +1,10 @@
 from django.utils import timezone
 
 from sentry.models import Group, GroupSnooze, GroupStatus
-from sentry.models.group import GroupSubStatus
 from sentry.models.grouphistory import GroupHistoryStatus, record_group_history
 from sentry.signals import issue_unignored
 from sentry.tasks.base import instrumented_task
+from sentry.types.group import GroupSubStatus
 
 
 @instrumented_task(name="sentry.tasks.clear_expired_snoozes", time_limit=65, soft_time_limit=60)
