+++ src/sentry/tasks/weekly_escalating_forecast.py
@@ -9,9 +9,9 @@ from sentry_sdk.crons.decorator import monitor
 from sentry.issues.escalating import GroupsCountResponse, query_groups_past_counts
 from sentry.issues.escalating_issues_alg import generate_issue_forecast
 from sentry.models import Group, GroupStatus
-from sentry.models.group import GroupSubStatus
 from sentry.models.groupforecast import GroupForecast
 from sentry.tasks.base import instrumented_task
+from sentry.types.group import GroupSubStatus
 
 BATCH_SIZE = 500
